http: replace io with fio module
In tarantool the standart built-in module for working with files is "fio".
So we should use it instead of "io".

Patch adds a commit 'Get rid of io module'
(e0d8f8266c98916dfce4c74c5b40adef6b714f70) that was implemented for http
v2 and later reverted in scope of issue with discard v2.

Follows up #90
Part of #134