http: disable display_errors option by default
When option 'display_errors' is enabled httpd return application errors
and backtraces to the client (like PHP). It is unexpected that the
debugging information is sent to a client by default. It is usual to
have such option, but have it disabled by default: see Django's DEBUG
setting [1] for example.

1. https://docs.djangoproject.com/en/3.2/ref/settings/#debug

Part of https://github.com/tarantool/security/issues/8