rpm: fix building rpm using packpack
Starting with fedora 33, the cmake build strategy has been changed
to "out-of-source builds" by default (see [1]). Let's add support
for the new behavior.

Added `prebuild.sh` script with installation of Tarantool repo.

`Packpack` performs building rpm packages in a clean docker container.
The container must contain `luatest` to perform testing during the
build as it was set by `-DCMAKE_BUILD_TYPE=RelWithDebInfo` cmake var.
This cmake var was removed in this commit as well as `%check` phase.

It's quite challenging to set up `luacheck` while the package is
building and a bit useless because the module is tested in another job.

[1] https://fedoraproject.org/wiki/Changes/CMake_to_do_out-of-source_builds

Closes #165