api: support `ssl_verify_client` option
This patch allows to set a new `ssl_verify_client` option. It uses in
pair with `ssl_ca_file` option and needs for client validation. It could
have following values:

* `off` (default one) means that no client's certs will be verified;
* `on` means that server will verify client's certs;
* `optional` means that server will verify client's certs only if it exist.

This set of options was was built on top of the NGINX API
(https://nginx.org/en/docs/http/ngx_http_ssl_module.html#ssl_verify_client).

In effect, this option forces the server to work with mutual TLS.

Part of #207